fix(lifecycle-updown): resolve verified lifecycle-updown sweep findings (#941)
I went through the verified lifecycle / up-down sweep findings and fixed
the ones that have a clean, well-scoped fix. The goal throughout was to
make `up`/`down`/`start`/`restart`/`rm`/`pause`/`unpause` behave the way
docker compose does: idempotent where it should be, best-effort across a
batch rather than aborting on the first failure, and loud about real
mistakes (unknown service names, conflicting flags, cycles) instead of
silently exiting 0.

Each behaviour change is covered by unit/parse-level tests (ordering
determinism, the flag conflict, config-time cycle detection, the new
error classifiers, and the unknown-service validation). The
container-touching paths stay covered by the existing integration suite.

## Fixed

- down --rmi no longer force-removes in-use images, which cascaded
container deletion across other projects sharing the image (Closes #711)
- resolve_order is deterministic, so wait's exit code and output order
are stable across runs (Closes #718)
- restart deduplicates a service that is both a target and a
restart-dependent so it is restarted once (Closes #751)
- restart is best-effort: it attempts every container and aggregates
errors instead of aborting mid-batch (Closes #752)
- up/create error on unknown/typo'd service names instead of a silent
exit-0 no-op (Closes #753)
- mutually-exclusive --no-recreate and --force-recreate are rejected
together (Closes #754)
- rm without -f skips running containers and continues, removing the
stopped ones (Closes #755)
- pause/unpause are idempotent and best-effort across the batch (Closes
#756)
- down's orphan sweep is gated on --remove-orphans so the flag is no
longer a no-op (Closes #864)
- start is best-effort: it aggregates errors instead of aborting on the
first (Closes #865)
- start warns "no containers to start" when the project was never
created, instead of a silent exit 0 (Closes #866)
- rm only logs "removed <name>" for containers that actually existed
(Closes #867)
- rm -s unpauses a paused container before stopping it (Closes #868)
- the spurious "cascade restart ... (HTTP 404)" warning on an empty
project is gone (Closes #869)
- down stays idempotently quiet on 404 for already-gone / profile-gated
services (Closes #871)
- config detects a depends_on cycle instead of only catching it at up
time (Closes #872)

## Deferred

- #757 (intra-level parallelism for lifecycle ops) — a sizable
performance refactor that changes dispatch/ordering semantics and would
conflict broadly with the other in-flight branches; better done on its
own with care.
- #870 (label-only down -p with no compose file) — requires
restructuring the parse-before-dispatch startup flow in main.rs; high
conflict risk, worth a dedicated PR.
- #873 (per-container progress/summary output) — a broad UX feature
across every command; out of scope for this correctness sweep.

---------
