+++ internal/compose/order.rs
@@ -1,4 +1,5 @@
-use std::collections::HashMap;
+use std::cmp::Reverse;
+use std::collections::{BinaryHeap, HashMap};
 
 use crate::compose::types::ComposeFile;
 use crate::error::{ComposeError, Result};
@@ -30,25 +31,27 @@ pub fn resolve_order(file: &ComposeFile) -> Result<Vec<String>> {
 		}
 	}
 
-	// Seed the queue from `services` (compose-file order) rather than iterating
-	// the `in_degree` HashMap, whose order is randomised per run. This keeps the
-	// resolved order deterministic for independent services, so dependent
-	// commands (e.g. best-effort `kill`) behave reproducibly.
-	let mut queue: std::collections::VecDeque<&str> = services
+	// A min-heap (lexicographically smallest name first) keeps the order
+	// deterministic: the in-degree map is a `HashMap`, so seeding/extending the
+	// frontier from its iteration order would otherwise be per-run random. This
+	// mirrors the per-level `sort_unstable` in `resolve_levels`, so independent
+	// (in-degree-0) services resolve in a stable order — which `wait` relies on
+	// for a reproducible exit code and output ordering.
+	let mut queue: BinaryHeap<Reverse<&str>> = in_degree
 		.iter()
-		.copied()
-		.filter(|s| in_degree.get(s) == Some(&0))
+		.filter(|(_, &deg)| deg == 0)
+		.map(|(&s, _)| Reverse(s))
 		.collect();
 
 	let mut order = Vec::new();
-	while let Some(node) = queue.pop_front() {
+	while let Some(Reverse(node)) = queue.pop() {
 		order.push(node.to_string());
 		let neighbors: Vec<&str> = graph.get(node).map_or(&[][..], |v| v.as_slice()).to_vec();
 		for neighbor in neighbors {
 			if let Some(deg) = in_degree.get_mut(neighbor) {
 				*deg -= 1;
 				if *deg == 0 {
-					queue.push_back(neighbor);
+					queue.push(Reverse(neighbor));
 				}
 			}
 		}
@@ -147,6 +150,37 @@ mod tests {
 	}
 
 	#[test]
+	fn resolve_order_is_deterministic_for_independent_services() {
+		// Independent (in-degree-0) services must resolve in a stable,
+		// lexicographic order regardless of the HashMap iteration order, so
+		// `wait`'s exit code and printed order are reproducible across runs.
+		let yaml = "services:\n  c:\n    image: x\n  a:\n    image: y\n  b:\n    image: z\n";
+		let file = parse_str_raw(yaml).unwrap();
+		let order = resolve_order(&file).unwrap();
+		assert_eq!(
+			order,
+			vec!["a".to_string(), "b".to_string(), "c".to_string()]
+		);
+		// Re-resolving yields the identical order.
+		for _ in 0..16 {
+			assert_eq!(resolve_order(&file).unwrap(), order);
+		}
+	}
+
+	#[test]
+	fn resolve_order_dependents_are_deterministic() {
+		// Two dependents of the same dependency come out in stable lexicographic
+		// order, not whatever the graph's adjacency iteration happens to be.
+		let yaml = "services:\n  db:\n    image: x\n  zeb:\n    image: y\n    depends_on: [db]\n  api:\n    image: z\n    depends_on: [db]\n";
+		let file = parse_str_raw(yaml).unwrap();
+		let order = resolve_order(&file).unwrap();
+		assert_eq!(
+			order,
+			vec!["db".to_string(), "api".to_string(), "zeb".to_string()]
+		);
+	}
+
+	#[test]
 	fn resolve_order_dep_before_dependent() {
 		let yaml = "services:\n  web:\n    image: nginx\n    depends_on: [db]\n  db:\n    image: postgres\n";
 		let file = parse_str_raw(yaml).unwrap();
@@ -154,23 +188,6 @@ mod tests {
 		let db_pos = order.iter().position(|s| s == "db").unwrap();
 		let web_pos = order.iter().position(|s| s == "web").unwrap();
 		assert!(db_pos < web_pos, "db must start before web");
-	}
-
-	#[test]
-	fn resolve_order_is_deterministic_for_independent_services() {
-		// Independent services must resolve in a stable (compose-file) order on
-		// every call, so best-effort consumers like `kill` behave reproducibly
-		// rather than depending on HashMap iteration order.
-		let yaml = "services:\n  a:\n    image: x\n  b:\n    image: y\n  c:\n    image: z\n";
-		let file = parse_str_raw(yaml).unwrap();
-		let first = resolve_order(&file).unwrap();
-		assert_eq!(
-			first,
-			vec!["a".to_string(), "b".to_string(), "c".to_string()]
-		);
-		for _ in 0..16 {
-			assert_eq!(resolve_order(&file).unwrap(), first);
-		}
 	}
 
 	#[test]
